requirements.txt
# AI知识库项目依赖

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1

# 搜索引擎
elasticsearch==8.11.0

# 向量数据库
qdrant-client==1.7.0

# AI服务
openai==1.3.7
httpx==0.25.2

# 文件处理
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
Pillow==10.1.0

# 文本处理
jieba==0.42.1
langdetect==1.0.9

# 工具库
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
loguru==0.7.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# 部署
gunicorn==21.2.0